Al menos sale

+++ Ej2_IoC/src/main/java/com/cris/IoC/almacenamiento/Almacenamiento.java
@@ -2,5 +2,5 @@ package com.cris.IoC.almacenamiento;
 
 public interface Almacenamiento {
 	
-	public Integer getIVA();
+	public float getIVA();
 }
